+++ server/api/server.js
@@ -1,0 +1,8 @@
+const express = require('express');
+
+const app = express();
+
+module.exports = {
+  app,
+  express,
+};
